+++ Changelog.md
@@ -16,6 +16,7 @@ Version 1.2
   * Press space when on a selected track also does quicklook.
   * Quicklook toolbar button.
 * Label colours for files shown.
+* Nicer display of label colour for selected row
 
 ### Improvements ###
 * Disable Open Selected Fodler if the user has not selected a folder.